form pemasukan

+++ database/migrations/2020_06_15_162021_create_pemasukkans_tables.php
@@ -16,7 +16,7 @@ class CreatePemasukkansTables extends Migration
         Schema::create('pemasukkans', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('total');
+            $table->decimal('total');
             $table->timestamps();
         });
     }
